requirements/base.txt
django-cms>=3.0c1
